V0.1.1 dev

+++ mirte_robot/linetrace.py
@@ -16,11 +16,19 @@ from mirte_robot import robot
 # Global shared memory objects (TODO: check if we need shared memory, why is server working?)
 stepper = multiprocessing.Value('b', True)
 do_step = multiprocessing.Value('b', False)
+running = multiprocessing.Value('b', False)
 
-def stop_mirte():
-     process.terminate()
+# the stop be stopped when:
+# 1) the code finishes
+# 2) the user stopped the process
+# 3) the websocket connection is closed
+def stop_mirte(terminate = True):
+     global running
+     if terminate:
+        process.terminate()
+     running.value = False
 
-def load_mirte_module(stepper, do_step):
+def load_mirte_module(stepper, do_step, running):
 
     def trace_lines(frame, event, arg):
        global do_step
@@ -43,7 +51,7 @@ def load_mirte_module(stepper, do_step):
     # Send the PID to the web interface and give it some time to call strace on this process
     # to see the output of this script
     server.send_message_to_all("pid:" + str(os.getpid()))
-    time.sleep(0.1)
+    time.sleep(0.2) #TODO: let client send signal when strace is started
 
     sys.settrace(traceit)
     # rospy.init_node() for some reason needs to be called from __main__ when importing in the regular way.
@@ -61,24 +69,32 @@ def load_mirte_module(stepper, do_step):
     # Sending the linetrace 0 to the client
     server.send_message_to_all("0")
 
+    stop_mirte(False)
+
 process = multiprocessing.Process(target = load_mirte_module, args=(stepper, do_step))
 
 def start_mirte():
     global process
-    process = multiprocessing.Process(target = load_mirte_module, args=(stepper, do_step))
+    # process should already have been killed after stop, or disconnect
+    # but, just in case make sure to stop this
+    if process.is_alive():
+      process.terminate()
+    process = multiprocessing.Process(target = load_mirte_module, args=(stepper, do_step, running))
     process.start()
 
 def client_left(client, server):
     stop_mirte()
 
 def message_received(client, server, message):
-   global stepper, do_step
+   global stepper, do_step, running
+
    if message == "b": #break (pause)
       stepper.value = True
    if message == "c": #continue (play)
       stepper.value = False
-      if not process.is_alive():
+      if not running.value:
          start_mirte()
+         running.value = True
    if message == "s": #step (step)
       do_step.value = True
    if message == "e": #exit (stop)
